hotel managment dashboard front & small backend fixes

+++ routes/systemAdmin/hotel.php
@@ -2,13 +2,13 @@
     use App\Http\Controllers\HotelController;
     use App\Http\Middleware\SystemAdminMiddleware;
 
-    Route::prefix('system')->controller(HotelController::class)->group(function () {
-            Route::prefix('hotel')->group(function() {
+    Route::prefix('admin')->controller(HotelController::class)->group(function () {
+            Route::prefix('hotels')->group(function() {
                 Route::middleware([SystemAdminMiddleware::class])->group(function () {
-                    Route::get('/index', 'adminIndex');
-                    Route::post('/create', 'store');
-                    Route::post('{hotel}/edit','update');
-                    Route::delete('{hotel}/delete' , 'destroy');
+                    Route::get('/', 'adminIndex')->name('admin.hotel.index');
+                    Route::post('/create', 'store')->name('admin.hotel.store');
+                    Route::put('{hotel}/edit','update')->name('admin.hotel.update');
+                    Route::delete('{hotel}/delete' , 'destroy')->name('admin.hotel.delete');
             });
         });
     });
